+++ dev/Menu.tsx
@@ -19,6 +19,7 @@ export const Menu: FC = () => {
           <li><Link to="/tailwind/navigation/nav-pills">NavPills</Link></li>
           <li><Link to="/tailwind/feedback/dialogs">Dialogs</Link></li>
           <li><Link to="/tailwind/feedback/result">Result</Link></li>
+          <li><Link to="/tailwind/feedback/message">Message</Link></li>
         </>
       ) : (
         <>